@@ -1,11 +1,14 @@
 <script>
-  export let activeCategory; // Get active category from the webinar/contouring page
   import { goto } from '$app/navigation';
-  export let currentPage; 
+  
+  let { 
+    activeCategory,
+    currentPage
+  } = $props();
 </script>
 
 <form method="get" action="/{currentPage}" id="filter" name="filter">
-  <button type="button" on:click={() => goto(`/${currentPage}`)} class:selected={!activeCategory || activeCategory === 'all'}>
+  <button type="button" onclick={() => goto(`/${currentPage}`)} class:selected={!activeCategory || activeCategory === 'all'}>
     All
   </button>
 
@@ -50,6 +53,13 @@
     -webkit-mask: linear-gradient(90deg,#0000,#000 5% 95%,#0000);
     mask: linear-gradient(90deg,#0000,#000 5% 95%,#0000);
     margin: 4em 0;
+
+    @media (min-width:600px) {
+      flex-wrap: wrap;
+      justify-content: center;
+      -webkit-mask: 0;
+      mask: 0;
+    }
   }
 
   button {
@@ -64,6 +74,11 @@
     color: inherit;
     transition: .2s;
     font-weight: 700;
+
+    @media (min-width:600px) {
+      margin-inline: 5px 8px;
+      padding: 8px 16px;
+    }
   }
   
   button:hover {
@@ -79,19 +94,5 @@
   button:focus {
     outline: var(--focus);
     border-radius: var(--border-radius-sm);
-  }
-
-  @media only screen and (min-width:600px) {
-    form {
-      flex-wrap: wrap;
-      justify-content: center;
-      -webkit-mask: 0;
-      mask: 0;
-    }
-
-    button {
-      margin-inline: 5px 8px;
-      padding: 8px 16px;
-    }
   }
 </style>
