@@ -1,7 +1,10 @@
 <script>
   import { PreviewDoc } from "$lib/index.js";
-  export let heading
-  export let resources
+  
+  let { 
+    heading,
+    resources
+  } = $props();
 </script>
 
 <section>
@@ -115,6 +118,10 @@
     background-color: var(--primary-color);
     margin-right: .1rem;
     transition: .2s;
+
+    @media (min-width: 370px) {
+      margin-left: .3rem;
+    }
   }
 
   .resource-links a:hover {
@@ -123,11 +130,5 @@
   
   .resource-links a path {
     fill: var(--alt-text-color);
-  }
-
-  @media only screen and (min-width: 370px) {
-    .resource-links a {
-      margin-left: .3rem;
-    }
   }
 </style>
