@@ -1,22 +1,27 @@
 <script>
+  import { run } from 'svelte/legacy';
   import { Navigation, Footer, LoadingState } from "$lib/index.js";
   import { navigating } from '$app/stores';
   import { onNavigate } from "$app/navigation";
 
-  let showLoading = false;
-  let navigatingTimeout;
+  let { children } = $props();
+
+  let showLoading = $state(false);
+  let navigatingTimeout = $state();
   
   // Watch the `navigating` store
-  $: if ($navigating) {
-    // Delay showing the loading state
-    clearTimeout(navigatingTimeout);
-    navigatingTimeout = setTimeout(() => {
-      showLoading = true;
-    }, 400); // Adjust delay for view-transition capture
-  } else {
-    clearTimeout(navigatingTimeout);
-    showLoading = false; // Hide loading when navigation completes
-  }
+  run(() => {
+    if ($navigating) {
+      // Delay showing the loading state
+      clearTimeout(navigatingTimeout);
+      navigatingTimeout = setTimeout(() => {
+        showLoading = true;
+      }, 400); // Adjust delay for view-transition capture
+    } else {
+      clearTimeout(navigatingTimeout);
+      showLoading = false; // Hide loading when navigation completes
+    }
+  });
   
   onNavigate((navigation) => {
     if (!document.startViewTransition) return;
@@ -38,7 +43,7 @@
 {:else} 
   <Navigation />
   <div class="content">
-    <slot />
+    {@render children?.()}
   </div>
 {/if}
 
@@ -53,32 +58,27 @@
   
   .content {
     flex: 1;
-  }
 
-  .footer {
-    background-color: var(--background-color);
-  }
-
-  @media only screen and (min-width: 600px) {
-    .content {
+    @media (min-width: 600px) {
       margin-bottom: 77px;
     }
-  }
-
-  @media only screen and (min-width: 1080px) {
-    .content {
+    
+    @media (min-width: 1080px) {
       margin-top: 74px;
       padding-bottom: 0;
     }
 
-    .footer {
+    @media (min-width: 1920px) {
+      margin-top: 82px;
+    }
+  }
+
+  .footer {
+    background-color: var(--background-color);
+
+    @media (min-width: 1080px) {
       background-color: #f0f0f0;
     }
   }
 
-  @media only screen and (min-width: 1920px) {
-    .content {
-      margin-top: 82px;
-    }
-  }
 </style>
