@@ -1,89 +1,55 @@
 <footer>
-  <div>
-    <p>&copy; 2024 - Antoni van Leeuwenhoek</p>
-    <ul class="dot-divider">
-      <li>
-        <a href="/">Disclaimer</a>
-      </li>
-      <li>
-        <a href="/">Privacystatement</a>
-      </li>
-      <li>
-        <a href="/">Cookiestatement</a>
-      </li>
-    </ul>
-    <div class="footer-logos">
-      <img src="/images/AVL-Logo-white.svg" alt="Antoni van Leeuwenhoek">
-      <img src="/images/kemenkessardjito.png" alt="kemenkes ziekenhuis">
-    </div>
+  <p>&copy; 2024 - Antoni van Leeuwenhoek</p>
+  <ul class="dot-divider">
+    <li>
+      <a href="/">Disclaimer</a>
+    </li>
+    <li>
+      <a href="/">Privacystatement</a>
+    </li>
+    <li>
+      <a href="/">Cookiestatement</a>
+    </li>
+  </ul>
+  <div class="footer-logos">
+    <img src="images/LogoOncollaboration.png" width="40" height="45" alt="Logo Antoni van Leeuwenhoek Hospital" class="s-lCPQtr6yAWsI">
+    <img src="images/Kemenkes2.png" width="40" height="45" alt="Logo Kemenkes hospital" class="s-lCPQtr6yAWsI">
   </div>
 </footer>
 
 <style>
   footer {
-    display: none;
-  }
+    height: 12.5em;
+    color: var(--text-light);
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-evenly;
+    padding-top: 0.6rem;
+    background-color: var(--primary-color);
 
-  @media only screen and (min-width: 1080px) {
-    footer {
-      display: flex;
+    @media (min-width: 1080px) {
       position: sticky;
       bottom: 0;
-      align-items: center;
-      justify-content: space-between;
-      background-color: var(--primary-color);
-      padding: 1rem 0rem;
-      color: var(--alt-text-color);
     }
+  }
 
-    footer ul {
-      display: flex;
-    }
-    
-    footer li {
-      padding: 0 0.8rem;
-      position: relative;
-    }
-    
-    footer div {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      margin: 0 auto;
-      width: 85vw;
-    }
+  footer ul {
+    display: flex;
+    justify-content: space-evenly;
+    width: 100%;
+  }
 
-    footer img {
-      height: 3.2rem;
-    }
+  footer li {
+    list-style: none;
+    font-size: 11px;
+  }
 
-    .footer-logos {
-      display: flex;
-      justify-content: flex-end;
-      align-items: center;
-      gap: 1rem; 
-    }
+  footer a {
+    color: var(--text-light);
+  }
 
-    footer a {
-      font-weight: normal;
-      color: var(--alt-text-color);
-      text-decoration: none;
-    }
-
-    footer a:hover {
-      text-decoration: underline;
-    }
-    
-    .dot-divider li::after {
-      content: '•';
-      position: absolute;
-      right: -0.2rem;
-      top: 50%;
-      transform: translateY(-50%);
-    }
-
-    .dot-divider li:last-child::after {
-      content: '';
-    }
+  footer a:hover {
+    text-decoration: underline;
   }
 </style>
