@@ -3,12 +3,15 @@ import Search from "./Blocks/search.svelte";
 import Filter from "./Blocks/filter.svelte";
 import WebinarOverview from "./Blocks/webinar-overview.svelte";
 import Navigation from "./Constructions/navigation.svelte";
+import Carrousel from "./Constructions/carrousel.svelte";
 import FeaturedWebinar from "./Blocks/featured-webinar.svelte";
 import Footer from "./Constructions/footer.svelte";
 import SearchResults from "./Constructions/search-results.svelte";
 import QandA from "./Assemblages/q-a.svelte";
 import Comment from "./Blocks/comment.svelte";
 import LoaderDots from "./Bricks/loader-dots.svelte";
+import FAQ from "./Blocks/FAQ.svelte";
+import Partners from "./Blocks/partners.svelte";
 import LoaderSpin from "./Bricks/loader-spin.svelte";
 import Like from "./Bricks/like.svelte";
 import Liked from "./Bricks/liked.svelte";
@@ -16,5 +19,10 @@ import LoadingState from "./Bricks/loading-state.svelte";
 import PreviewDoc from "./Bricks/preview-doc.svelte";
 import Resources from "./Blocks/resources.svelte";
 import EmptyState from "./Bricks/empty-states.svelte"
+import Article from "./Constructions/article.svelte";
+import Arrow from "./Bricks/arrow.svelte";
+import CTA from "./Constructions/cta.svelte";
+import DoctorsUnited from "./Constructions/doctors-united.svelte";
 
-export {ContouringOverview, Search, Filter, WebinarOverview, Navigation, FeaturedWebinar, Footer, SearchResults, QandA, Comment, LoaderDots, LoaderSpin, Like, Liked, LoadingState, PreviewDoc, Resources, EmptyState}
+export { ContouringOverview, Search, Filter, WebinarOverview, Navigation, Carrousel, FeaturedWebinar, Footer, SearchResults, QandA, Comment, LoaderDots, LoaderSpin, Like, Liked, LoadingState, PreviewDoc, Resources, EmptyState, Article, FAQ, Partners, Arrow, CTA, DoctorsUnited};
+
